chore: トーク詳細ページを /wip に移す
