+++ src/app/wip/talks/[id]/page.tsx
@@ -1,41 +1,47 @@
-import type { Metadata } from "next";
-import { talkList } from "../../../../constants/talkList";
+import { talkIds } from "@/constants/talkList";
+import { getTalk } from "@/utils/getTalk";
+import Link from "next/link";
+import type { ComponentProps } from "react";
+
+import Markdown from "react-markdown";
+import remarkBreaks from "remark-breaks";
 
 export async function generateStaticParams() {
-  return talkList.map((talk) => ({
-    id: talk.id,
-  }));
+  return talkIds;
 }
 
-export async function generateMetadata({
-  params,
-}: {
-  params: Promise<{ id: string }>;
-}): Promise<Metadata> {
-  const { id } = await params;
-  const talk = talkList.find((talk) => talk.id === id);
-  if (!talk) {
-    throw new Error("Talk not found");
-  }
-  return {
-    twitter: {
-      title: talk.title,
-      images: [
-        {
-          url: `/ogp/talks/${talk.id}.png`,
-        },
-      ],
-    },
-    openGraph: {
-      title: talk.title,
-      images: [
-        {
-          url: `/ogp/talks/${talk.id}.png`,
-        },
-      ],
-    },
-  };
-}
+const components: ComponentProps<typeof Markdown>["components"] = {
+  h1: ({ node, ...props }) => (
+    <h1 className="text-2xl font-bold text-blue-light-500" {...props} />
+  ),
+  h2: ({ node, ...props }) => (
+    <h2 className="text-xl font-bold text-blue-light-500" {...props} />
+  ),
+  h3: ({ node, ...props }) => (
+    <h3 className="text-lg font-bold text-blue-light-500" {...props} />
+  ),
+  a: ({ node, href, ...props }) => {
+    if (!href) return null;
+    return (
+      <Link
+        target="_blank"
+        rel="noopener noreferrer"
+        className="inline-flex items-center text-link-light hover:underline"
+        href={href}
+        {...props}
+      />
+    );
+  },
+  ul: ({ node, ...props }) => (
+    <ul className="text-gray-700 list-disc list-inside pl-6" {...props} />
+  ),
+  ol: ({ node, ...props }) => (
+    <ol className="text-gray-700 list-decimal list-inside pl-6" {...props} />
+  ),
+  pre: ({ node, ...props }) => (
+    <pre className="bg-gray-100 p-4 rounded-lg text-wrap" {...props} />
+  ),
+};
 
 export default async function TalkDetailPage({
   params,
@@ -43,10 +49,60 @@ export default async function TalkDetailPage({
   params: Promise<{ id: string }>;
 }) {
   const { id } = await params;
-  const talk = talkList.find((talk) => talk.id === id);
-  if (!talk) {
-    throw new Error("Talk not found");
-  }
+  const talk = getTalk(id);
 
-  return <main>talk</main>;
+  return (
+    <main className="bg-blue-light-100 pt-16 pb-10 md:py-16 md:px-8 lg:px-10">
+      <h1 className="text-2xl font-bold text-blue-light-500 text-center py-10 md:py-16 md:text-3xl lg:text-4xl">
+        トーク
+      </h1>
+
+      <div className="bg-white flex flex-col gap-6 max-w-screen-xl mx-auto md:rounded-xl pb-6 md:pb-8 lg:pb-10">
+        {/* トーク OGP */}
+        <div className="bg-black-100 flex justify-center md:mt-8 md:mx-8 lg:mt-10 lg:mx-10">
+          <img
+            width="730"
+            height="383"
+            className="w-full max-w-[730px] h-auto max-h-[383px] mx-auto object-contain"
+            src={`/ogp/talks/${talk.id}.png`}
+            alt="logo"
+          />
+        </div>
+
+        {/* トーク説明文 */}
+        <div className="px-6 md:px-8 lg:px-10 gap-6 flex flex-col md:text-lg">
+          <Markdown components={components} remarkPlugins={[remarkBreaks]}>
+            {talk.overview}
+          </Markdown>
+        </div>
+
+        {/* スピーカー情報 */}
+        {/* FIXME: デザイン未調整 */}
+        {/* <div className="pt-8 p-10">
+          <div className="bg-blue-light-200 p-6 rounded-xl">
+            <div className="flex flex-col md:flex-row items-center gap-4">
+              <div className="w-24 h-24 rounded-full overflow-hidden">
+                <img
+                  src={`/talks/speaker/${talk.id}.jpg`}
+                  alt={talk.speakerName}
+                  className="w-full h-full object-cover"
+                />
+              </div>
+              <div>
+                <p className="font-bold text-lg">{talk.speakerName}</p>
+                <div className="flex gap-2 mt-2">
+                  <a href="#" target="_blank" rel="noopener noreferrer">
+                    <TwitterIcon />
+                  </a>
+                  <a href="#" target="_blank" rel="noopener noreferrer">
+                    <GithubIcon />
+                  </a>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div> */}
+      </div>
+    </main>
+  );
 }
